# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Reimbursement Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /companies/{id}/currency:
    patch:
      operationId: updateCompanyCurrency
      tags: [companies]
      summary: Update company default currency
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCurrencyRequest"
      responses:
        "200":
          description: Currency updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users:
    get:
      operationId: getUsers
      tags: [users]
      summary: Get all users in company (admin only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a new user (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/{id}:
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user role or manager (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Company:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        country:
          type: string
        defaultCurrency:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - country
        - defaultCurrency
        - createdAt
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, manager, employee]
        managerId:
          type: ["integer", "null"]
        companyId:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - companyId
        - createdAt
...
    CreateUserRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [manager, employee]
        managerId:
          type: ["integer", "null"]
      required:
        - name
        - email
        - password
        - role
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
          enum: [admin, manager, employee]
        managerId:
          type: ["integer", "null"]
    UpdateCurrencyRequest:
      type: object
      properties:
        currency:
          type: string
      required:
        - currency
...